Add special handling for OnePlus 6 & 6T
It seems like FBE decryption works most reliably if both
keymaster@3.0 and @4.0 service binaries exist, but only
the 4.0 service is called in the init.recovery.* rc scripts
on fajita (OnePlus 6T). Linking issue?

Adding pre-munged init.recovery.qcom_decrypt[.fbe].rc files
and $(TARGET_DEVICE) tests to makefile to get fajita working
for now.

The Big Brain solution would be to edit the bash script to test
for multiple versions of the same service, and have it only
leave the lines enabling/stopping that highest-numbered service
version in the rc files when it runs post-install.

I don't know how to do that, but it would allow things to work
without device-specific logic tests in the makefile.

Change-Id: Iaac64f382d5716e49a0fbbbf6ef2c90bcd9c4f27